+++ bench/irmin-pack/main.ml
@@ -21,6 +21,7 @@ module Config = struct
   let stable_hash = 3
   let contents_length_header = Some `Varint
   let inode_child_order = `Hash_bits
+  let forbid_empty_dir_persistence = false
 end
 
 module KV = struct