+++ src/irmin-pack/inode.ml
@@ -1944,6 +1944,10 @@ struct
     let hash_exn ?force t = apply t { f = (fun _ v -> I.hash_exn ?force v) }
 
     let save ~add ~index ~mem t =
+      if Conf.forbid_empty_dir_persistence && is_empty t then
+        failwith
+          "Persisting an empty node is forbidden by the configuration of the \
+           irmin-pack store";
       let f layout v =
         I.check_write_op_supported v;
         I.save layout ~add ~index ~mem v